Clean up config test

+++ lib/core/src/__test__/config.test.ts
@@ -17,48 +17,39 @@ type TestSchema = typeof testSchema.infer;
 
 describe("config", () => {
   const validConfigWithAuth = { mqtt: { url: "mqtt://localhost:1883", username: "user", password: "pass" } };
+  let configPath: string;
+
+  beforeEach(() => {
+    configPath = `${os.tmpdir()}/config.json`;
+  });
+
+  afterEach(() => {
+    if (fs.existsSync(configPath)) {
+      fs.unlinkSync(configPath);
+    }
+  });
 
   it("should parse config file and produce config object", () => {
-    const configPath = `${os.tmpdir()}/config.json`;
     fs.writeFileSync(configPath, JSON.stringify(validConfigWithAuth, null, 2));
-
-    // const { data, problems } = testSchema({});
-    //     ^?
 
     const { data, problems } = testSchema(parseConfig(configPath));
     expect(problems).toBeUndefined();
     expect(data).not.toBeUndefined();
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const result: TestSchema = data!;
+    const result = data!;
 
-    // expectTypeOf(result).toMatchTypeOf<BaseOutput>();
+    expectTypeOf(result).toMatchTypeOf<TestSchema>();
 
     expect(result).toStrictEqual(validConfigWithAuth);
-
-    fs.unlinkSync(configPath);
-  });
-
-  it("should throw error if config file is not valid", () => {
-    const configPath = `${os.tmpdir()}/config.json`;
-    fs.writeFileSync(configPath, JSON.stringify({ quux: { url: "mqtt://localhost:1883" } }, null, 2));
-
-    // expect(() => parseConfig(configPath, BaseTestSchema)).toThrowError(/Invalid config file/);
-
-    fs.unlinkSync(configPath);
   });
 
   it("should throw error if config file does not exist", () => {
-    // expect(() => parseConfig("does-not-exist.json", BaseTestSchema)).toThrowError(
-    //   /Could not read config file/
-    // );
+    expect(() => parseConfig("does-not-exist.json")).toThrowError(/Could not read config file/);
   });
 
   it("should throw error if config file is not valid JSON", () => {
-    const configPath = `${os.tmpdir()}/config.json`;
     fs.writeFileSync(configPath, "this is not valid JSON");
 
-    // expect(() => parseConfig(configPath, BaseTestSchema)).toThrowError(/Could not parse config file/);
-
-    fs.unlinkSync(configPath);
+    expect(() => parseConfig(configPath)).toThrowError(/Could not parse config file/);
   });
 });
